pyproject.toml
[tool.poetry]
name = "lingyin-bot"
version = "0.1.0"
description = ""
authors = ["chrisyy <1017975501@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0rc2"
nonebot-adapter-onebot = "^2.1.5"
nonebot-plugin-apscheduler = "^0.2.0"
revchatgpt = "0.0.34"
poetry = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# nonebot配置
[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["plugins/"]
